move doc gen code to gen/ package
