+++ pkg/tools/gen/templates/doc/schemaDoc.gotmpl
