+++ pkg/tools/gen/templates/kcl/document.gotmpl
