+++ pkg/tools/gen/templates/kcl/schema.gotmpl
