+++ pkg/tools/gen/templates/kcl/validator.gotmpl
