+++ pkg/tools/gen/testdata/doc/pkg/server.k
